% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.R
\name{area}
\alias{area}
\title{Triangle area}
\usage{
area(x)
}
\arguments{
\item{x}{A matrix containing the coordinates of the triangle. Each row
corresponds to a points, and each column to a dimension (x, y and z).}
}
\description{
...
}

