% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canis.R
\docType{package}
\name{canis}
\alias{canis}
\alias{canis-package}
\title{canis: Canopy File Investigator and Shreder}
\description{
Toolkit to plot \code{.can} files in 3D, and detect inclusions of polygons in
a set of volumes.
}
\details{
Note that currently, \code{canis} can only deal with \code{.can} files
describing triangles. In addition, each volume generated in the set of
volumes must be a rectangular parallelepiped with sides parallel to x, y, or
z axes.
}

