% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{canopy3d}
\alias{canopy3d}
\title{3D canopy visualization}
\usage{
canopy3d(data, fraction = 1, color, ...)
}
\arguments{
\item{data}{A data frame containing canopy data.}

\item{fraction}{Which fraction of the data should be displayed. This is
useful when you deal with a huge amount of polygon in order to render
them in a reasonible amount of time.}
}
\description{
3D canopies visualized using rgl package features.
}

