% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.R
\name{containCanopy}
\alias{containCanopy}
\title{Interection between canopy polygons and a cube}
\usage{
containCanopy(can, volume, shred = TRUE, epsilon = 1e-06)
}
\description{
Check if each triangle is inside or outside the volume.
Shred a triangle into four smaller triangles
}
\details{
The "cube" should have all its sides parallel to an origin axis (current
restriction).

#@param x A matrix containing the coordinates of the triangle. Each row
    corresponds to a points, and each column to a dimension (x, y and z).
#@param volume A 3-component matrix of coordinates of a "cube".
#@param epsilon The area limit below which no triangle is to be shreded
    anymore.

#@export
}

