% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten}
\alias{flatten}
\alias{flatten.Canopy}
\title{Flatten a data frame with list-columns}
\usage{
flatten(data, ...)

\method{flatten}{Canopy}(data, ...)
}
\arguments{
\item{data}{The data frame to be flattened.}

\item{...}{Not currently used.}
}
\description{
Flattens a data frame (or an inherited object, such as \code{Canopy}) with
list-columns in order to get a regular data frame. This is for instance required
to export a data frame in a file.
}
\examples{
flatPlants <- flatten(plants)
filename <- tempfile(fileext = ".csv")
write.csv(flatPlants, filename, row.names = FALSE)

}

