% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.R
\name{heron}
\alias{heron}
\title{Heron's formula}
\usage{
heron(x)
}
\arguments{
\item{x}{A vector containing the lengths of the triangle sides.}
}
\description{
Compute the area of a triangle knowing only the lengths of its three sides.
}

