% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{readCan}
\alias{readCan}
\title{Read a canopy file (.can)}
\usage{
readCan(filePath)
}
\arguments{
\item{filePath}{The name of the file which the data are to be read from.}
}
\description{
Read .can files from .can files and save the information in a data.frame.
}

