% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{splitCanopy}
\alias{splitCanopy}
\title{Split a whole canopy in accordance with a set of volumes}
\usage{
splitCanopy(canopy, volumeSet, shred = TRUE, epsilon = 1e-06,
  parallel = FALSE, nCores = 2)
}
\arguments{
\item{canopy}{A \code{Canopy} object.}

\item{volumeSet}{A \code{VolumeSet} object.}

\item{shred}{Logical. Should an intersected polygon be shreded into smaller
polygons?}

\item{epsilon}{Minimal polygon area to take into account.}

\item{parallel}{Logical. Parallelization activation.}

\item{nCores}{Number of cores to use if \code{parallel} is set to \code{TRUE}.}
}
\value{
A \code{c("Canopy", "data.frame")} object containing only the
polygons that were successfully allocated to a volume.
}
\description{
Each polygon of the canopy that is included in one of the provided volumes is
allocated to this volume. If a polygon crosses one or several sides of a
volume and if the parameter \code{shred} is set to \code{TRUE}, then the
polygon is recursively shreded into smaller polygons until none of them (with
an area > \code{epsilon}) crosses the volume sides anymore.
}
\examples{
rangeDim <- function(can, dim) {
    range(sapply(can$vertices, function(i) range(i[, dim])))
}
volumes <- makeVolumeSet(rangeDim(plants, "x"),
                         rangeDim(plants, "y"),
                         rangeDim(plants, "z"),
                         intervals = rep(3, 3))
\donttest{
# With no parallelization:
allocatedPolygons <- splitCanopy(plants, volumes, shred = TRUE)

# With parallelization (experimental feature):
allocatedPolygons <- splitCanopy(plants, volumes, shred = TRUE,
                                 parallel = TRUE, nCores = 8)
}

}

