% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{readCan}
\alias{readCan}
\title{Read a canopy file (.can)}
\usage{
readCan(filePath)
}
\arguments{
\item{filePath}{The name of the file which the data are to be read from.}
}
\description{
Read .can files from .can files and save the information in a data.frame.
}
\details{
The data are stored in a data frame where the first columns correspond to
ID and characteristics of the polygons. The last columns contains the points
coordinates, one column per point. The list-columns in data frames.
}

