% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy.R
\name{readCan}
\alias{readCan}
\title{Data input: canopy file (.can)}
\usage{
readCan(file, label = NA, col3Name = c("speciesID", "plantID", "leafID",
  "polygonID"), col3Sep = c(12, 7, 4))
}
\arguments{
\item{file}{The name of the can file which the data are to be read from.}

\item{label}{Not yet implemeted.}

\item{col3Name}{Vector containing the names of the different elements in the
third columne of the can file.}

\item{col3Sep}{Positions to split at in the third column. It is mandatory
that \code{length(col3Sep)} equals \code{length(col3Name) - 1}.}
}
\value{
An object of class \code{c("data.frame", "Canopy")}.
}
\description{
Reads a canopy file, with a .can extension, and creates a data frame from it.
}
\details{
The data are stored in a data frame where each row corresponds to only one
constituent polygon of the canopy. Only triangle shapes can currently be
imported. The first columns contain "labels" linked to the polygons, i.e.
characteristics (other than the shape) and IDs. The last column is a
list-column containing all the vertices of the polygons. Each element of this
column is a matrix. Each row of this matrix corresponds to the spatial
coordinates of a vertex.
}
\examples{
plants <- readCan("/path/to/file/field.can")

}

